pyproject.toml
[tool.poetry]
name = "csaatVideoUpload"
version = "0.1.0"
description = ""
authors = ["sachinAthu <sachin2262716@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
PyMySQL = "^1.0.2"
Pillow = "^8.1.0"
Faker = "^5.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
